ext: move CustomExtension into the ext module
Move the type next to its `Extension` impl, adding whitespace between
the impl's methods; the crate root re-export path is unchanged.
