+++ rcgen/src/certificate.rs
@@ -11,9 +11,8 @@ use yasna::Tag;
 use crate::crl::CrlDistributionPoint;
 use crate::csr::CertificateSigningRequest;
 use crate::ext::{
-	AuthorityKeyIdentifier, BasicConstraints, Criticality, CrlDistributionPoints, ExtendedKeyUsage,
-	Extensions, KeyUsage, NameConstraints as NameConstraintsExt, SubjectAlternativeName,
-	SubjectKeyIdentifier,
+	AuthorityKeyIdentifier, BasicConstraints, CrlDistributionPoints, ExtendedKeyUsage, Extensions,
+	KeyUsage, NameConstraints as NameConstraintsExt, SubjectAlternativeName, SubjectKeyIdentifier,
 };
 use crate::key_pair::{serialize_public_key_der, sign_der, PublicKeyData};
 #[cfg(feature = "crypto")]
@@ -21,8 +20,8 @@ use crate::ring_like::digest;
 #[cfg(feature = "pem")]
 use crate::ENCODE_CONFIG;
 use crate::{
-	oid, write_distinguished_name, write_dt_utc_or_generalized, DistinguishedName, Error, Issuer,
-	KeyIdMethod, KeyUsagePurpose, SanType, SerialNumber, SigningKey,
+	oid, write_distinguished_name, write_dt_utc_or_generalized, CustomExtension, DistinguishedName,
+	Error, Issuer, KeyIdMethod, KeyUsagePurpose, SanType, SerialNumber, SigningKey,
 };
 
 /// An issued certificate
@@ -477,59 +476,6 @@ pub struct Attribute {
 	///
 	/// [RFC 2986]: https://datatracker.ietf.org/doc/html/rfc2986#section-4
 	pub values: Vec<u8>,
-}
-
-/// A custom extension of a certificate, as specified in
-/// [RFC 5280](https://tools.ietf.org/html/rfc5280#section-4.2)
-#[derive(Debug, PartialEq, Eq, Hash, Clone)]
-pub struct CustomExtension {
-	pub(crate) oid: Vec<u64>,
-	pub(crate) criticality: Criticality,
-
-	/// The content must be DER-encoded
-	pub(crate) content: Vec<u8>,
-}
-
-impl CustomExtension {
-	/// Creates a new acmeIdentifier extension for ACME TLS-ALPN-01
-	/// as specified in [RFC 8737](https://tools.ietf.org/html/rfc8737#section-3)
-	///
-	/// Panics if the passed `sha_digest` parameter doesn't hold 32 bytes (256 bits).
-	pub fn new_acme_identifier(sha_digest: &[u8]) -> Self {
-		assert_eq!(sha_digest.len(), 32, "wrong size of sha_digest");
-		let content = yasna::construct_der(|writer| {
-			writer.write_bytes(sha_digest);
-		});
-		Self {
-			oid: oid::PE_ACME.to_owned(),
-			criticality: Criticality::Critical,
-			content,
-		}
-	}
-	/// Create a new custom extension with the specified content
-	pub fn from_oid_content(oid: &[u64], content: Vec<u8>) -> Self {
-		Self {
-			oid: oid.to_owned(),
-			criticality: Criticality::NonCritical,
-			content,
-		}
-	}
-	/// Sets the criticality flag of the extension.
-	pub fn set_criticality(&mut self, criticality: bool) {
-		self.criticality = criticality.into();
-	}
-	/// Obtains the criticality flag of the extension.
-	pub fn criticality(&self) -> bool {
-		self.criticality == Criticality::Critical
-	}
-	/// Obtains the content of the extension.
-	pub fn content(&self) -> &[u8] {
-		&self.content
-	}
-	/// Obtains the OID components of the extensions, as u64 pieces
-	pub fn oid_components(&self) -> impl Iterator<Item = u64> + '_ {
-		self.oid.iter().copied()
-	}
 }
 
 #[derive(Debug, PartialEq, Eq, Hash, Clone)]
